docs: add missing copyright comments
