+++ src/Flicking.ts
@@ -1,8 +1,7 @@
-/**
+/*
  * Copyright (c) 2015 NAVER Corp.
  * egjs projects are licensed under the MIT license
  */
-
 import Component from "@egjs/component";
 import ImReady from "@egjs/imready";
 import { DiffResult } from "@egjs/list-differ";
