+++ src/camera/index.ts
@@ -1,3 +1,7 @@
+/*
+ * Copyright (c) 2015 NAVER Corp.
+ * egjs projects are licensed under the MIT license
+ */
 import Camera, { CameraOptions } from "./Camera";
 import LinearCamera from "./LinearCamera";
 import CircularCamera from "./CircularCamera";
