+++ src/const/external.ts
@@ -1,3 +1,7 @@
+/*
+ * Copyright (c) 2015 NAVER Corp.
+ * egjs projects are licensed under the MIT license
+ */
 /**
  * Event type object with event name strings.
  *
