--- src/const/internal.ts
@@ -1,15 +1,0 @@
-export const AXES_EVENT = {
-  HOLD: "hold",
-  CHANGE: "change",
-  RELEASE: "release",
-  ANIMATION_END: "animationEnd",
-  FINISH: "finish"
-} as const;
-
-export enum STATE_TYPE {
-  IDLE,
-  HOLDING,
-  DRAGGING,
-  ANIMATING,
-  DISABLED
-}
