+++ src/control/index.ts
@@ -1,3 +1,7 @@
+/*
+ * Copyright (c) 2015 NAVER Corp.
+ * egjs projects are licensed under the MIT license
+ */
 import Control from "./Control";
 import SnapControl, { SnapControlOptions } from "./SnapControl";
 import FreeControl from "./FreeControl";
