+++ src/control/states/IdleState.ts
@@ -1,10 +1,10 @@
-/**
+/*
  * Copyright (c) 2015 NAVER Corp.
  * egjs projects are licensed under the MIT license
  */
 import State from "~/control/states/State";
+import { STATE_TYPE } from "~/control/StateMachine";
 import { EVENTS } from "~/const/external";
-import { STATE_TYPE } from "~/const/internal";
 import { getDirection } from "~/utils";
 
 class IdleState extends State {
