+++ src/core/FlickingError.ts
@@ -1,3 +1,7 @@
+/*
+ * Copyright (c) 2015 NAVER Corp.
+ * egjs projects are licensed under the MIT license
+ */
 class FlickingError extends Error {
   public constructor(
     public message: string,
