+++ src/core/index.ts
@@ -1,7 +1,13 @@
+/*
+ * Copyright (c) 2015 NAVER Corp.
+ * egjs projects are licensed under the MIT license
+ */
 import Panel from "./Panel";
 import Viewport from "./Viewport";
+import FlickingError from "./FlickingError";
 
 export {
   Panel,
-  Viewport
+  Viewport,
+  FlickingError
 };
