+++ src/renderer/RawRenderer.ts
@@ -1,3 +1,7 @@
+/*
+ * Copyright (c) 2015 NAVER Corp.
+ * egjs projects are licensed under the MIT license
+ */
 import Renderer from "./Renderer";
 
 class RawRenderer extends Renderer {
