+++ src/renderer/index.ts
@@ -1,3 +1,7 @@
+/*
+ * Copyright (c) 2015 NAVER Corp.
+ * egjs projects are licensed under the MIT license
+ */
 import Renderer from "./Renderer";
 import RawRenderer from "./RawRenderer";
 import VisibleRenderer from "./VisibleRenderer";
