+++ src/types.ts
@@ -1,9 +1,8 @@
+/*
+* Copyright (c) 2015 NAVER Corp.
+* egjs projects are licensed under the MIT license
+*/
 /* eslint-disable @typescript-eslint/consistent-type-definitions */
-/**
- * Copyright (c) 2015 NAVER Corp.
- * egjs projects are licensed under the MIT license
- */
-
 import Component from "@egjs/component";
 import { DiffResult } from "@egjs/list-differ";
